Add create_table to Database class
I added a property to keep track of the metadata channel and make it
easier to retrieve data from it or write data to it.

create_table creates a table in the database and all index tables that
go with it. It then writes the metadata about the table to the metadata
channel